Fix build with later WDKs:
- Adds alias for GetProjectInfoForReference target to version.vcxproj
  Later kits seemed to have renamed the build target, and will fail without
  this alias target.
- Adds "/fd sha256" to signtool command line
  WDK 20344 and later require binaries signed with a SHA256 file digest, or
  the build outputs are deleted.


Re-worked from Owen's original patch:

- Squashes warnings 4061 and 26052.
- Casts XENBUS_STORE_PERMISSION_MASK to ULONG in switch statement to avoid
  complaint about case using '|'.

+++ src/xenbus/store.c
@@ -1945,7 +1945,7 @@ StorePermissionToString(
 
     ASSERT(BufferSize > 1);
 
-    switch (Permission->Mask) {
+    switch ((ULONG)Permission->Mask) {
     case XENBUS_STORE_PERM_NONE:
         *Buffer = 'n';
         break;
